requirements.txt
# IshemaLink — Python Dependencies

# Core
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
django-cors-headers==4.3.1

# Async / WebSocket
channels==4.1.0
channels-redis==4.2.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0

# Database
psycopg2-binary==2.9.9

# Cache
django-redis==5.4.0

# Celery (async tasks)
celery==5.4.0
redis==5.0.4

# OpenAPI
drf-spectacular==0.27.2

# Monitoring
django-prometheus==2.3.1
prometheus-client==0.20.0

# Logging
python-json-logger==2.0.7

# Testing
pytest==8.2.0
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0

# Load testing (separate install)
# locust==2.29.0

# Security scanning
bandit==1.7.8
safety==3.2.0
